security-intel-rag/requirements.txt
# Core dependencies
anthropic==0.39.0                # Claude API
python-dotenv==1.0.0            # Environment variable management

# Vector database and embeddings
chromadb==0.4.22                # Local vector database (free!)
sentence-transformers==2.3.1    # Free embedding models (no API needed)

# Data processing
pandas==2.1.4                   # Data manipulation
requests==2.31.0                # HTTP requests for data fetching
beautifulsoup4==4.12.3          # HTML parsing (for threat feeds)
lxml==5.1.0                     # XML parsing (for CVE data)

# API framework
fastapi==0.109.0                # Modern Python web framework
uvicorn==0.27.0                 # ASGI server
pydantic==2.5.3                 # Data validation

# Utilities
tqdm==4.66.1                    # Progress bars
python-dateutil==2.8.2          # Date handling

# Development
jupyter==1.0.0                  # For demo notebook
ipykernel==6.29.0              # Jupyter kernel
pytest==7.4.4                   # Testing framework
